% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{choinv}
\alias{choinv}
\title{Fast Matrix Inverse for Positive Definite Matrix by Cholesky Decomposition}
\usage{
choinv(X)
}
\arguments{
\item{X}{The positive definite matrix.}
}
\value{
The inverse of X.
}
\description{
Fast Matrix Inverse for Positive Definite Matrix by Cholesky Decomposition
}
\details{
Implemented by RcppEigen
}
