% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{crossprod_fast}
\alias{crossprod_fast}
\title{Fast Matrix Dot Product with Transpose to Replace "crossprod".}
\usage{
crossprod_fast(x, y = NULL)
}
\arguments{
\item{x}{The input matrix.}

\item{y}{The input matrix or vector or NULL.}
}
\value{
The dot product of the transpose of 'x' and 'y'.
}
\description{
Fast Matrix Dot Product with Transpose to Replace "crossprod".
}
\details{
Implemented by RcppEigen
}
