% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{expit}
\alias{expit}
\title{Expit Function from Rcpp}
\usage{
expit(x)
}
\arguments{
\item{x}{x can be a single number or a vector or a matrix.}
}
\value{
the element-wise expit value.
}
\description{
Expit Function from Rcpp
}
\details{
expit(s)=exp(s)/{1+exp(s)}
}
