% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{gwas.htlgmm.repeated.term}
\alias{gwas.htlgmm.repeated.term}
\title{Function to compute repeated term in gwas.htlgmm}
\usage{
gwas.htlgmm.repeated.term(y, A = NULL, family = "gaussian")
}
\arguments{
\item{y}{response variable}

\item{A}{covariates adjusted in external GWAS}

\item{family}{gaussian or binomial}
}
\value{
the repeated term list.
}
\description{
Function to compute repeated term in gwas.htlgmm
}
\details{
expit(s)=exp(s)/{1+exp(s)}
}
