% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{sqrtchoinv}
\alias{sqrtchoinv}
\title{Fast Matrix Square Root Inverse for Positive Definite Matrix by Cholesky Decomposition}
\usage{
sqrtchoinv(X)
}
\arguments{
\item{X}{The positive definite matrix.}
}
\value{
The inverse of X.
}
\description{
Fast Matrix Square Root Inverse for Positive Definite Matrix by Cholesky Decomposition
}
\details{
Implemented by RcppEigen
}
