% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{gwas.htlgmm}
\alias{gwas.htlgmm}
\title{GWAS version of htlgmm.}
\usage{
gwas.htlgmm(
  y,
  Z,
  W = NULL,
  ext_study_info = NULL,
  A = "default",
  family = "gaussian",
  beta_initial = NULL,
  repeated_term = NULL,
  refine_C = FALSE,
  output_SNP_only = TRUE,
  seed.use = 97,
  verbose = FALSE,
  output_tmp = FALSE
)
}
\arguments{
\item{y}{The variable of interest, which can be continuous or binary.}

\item{Z}{The overlapping features in both main and external studies.}

\item{W}{The unmatched features only in main study, the default is NULL.}

\item{ext_study_info}{The trained model from external GWAS study, including estimated coefficients, and estimated variance.
The 'ext_study_info' is in the format of list. The first item is 'Coeff', the second item is 'Covariance'.
E.g. ext_study_info = list(list("Coeff"=coeff,"Covariance"=variance))}

\item{A}{The covariates for study-specific adjustment. The default is 'default', which is 'NULL' for 'gaussian' family, '1' for 'binomial' family.
Other than c('default',NULL,1), A must be a matrix whose dimension is the same as the sample dimension or Z and W.
For continuous variable, we suggest scaling the features Z, W to eliminate intercept term.  If 'A = NULL', there is no intercept term included.
For binary variable, we use intercept term by 'A=1' to adjust for different binary trait ratios in main and external studies.
If there is only intercept term in A, we use 'A=1'.
A are the features working for adjustment in reduced model, but A is not summarized in summary statistics(input:ext_study_info).}

\item{family}{The family is chosen from c("gaussian","binomial"). Linear regression for "gaussian" and logistic regression for "binomial".}

\item{beta_initial}{The beta_initial list matching ext_study_info, which provides the initial value of beta, e.g. fitting full model using plink2. If there is only one SNP, the beta_initial can be a vector.
The default is NULL. The beta_initial should go in the order of (A,W,Z), where Z is for each SNP.}

\item{repeated_term}{Default is NULL, which includes A_hat_thetaA, V_thetaA, and inv_GammaAA.}

\item{refine_C}{When computing the variance, whether recompute the weighting matrix C using final estimated beta.}

\item{output_SNP_only}{Default is TRUE.}

\item{seed.use}{The seed for  97.}

\item{verbose}{Default is FALSE.}

\item{output_tmp}{Default is FALSE}
}
\value{
\itemize{
 \item{beta:} The target coefficient estimation, the features will go in the order of (A,Z,W).
 \item{variance:} The lambda list for cross validation.
 }
}
\description{
GWAS version of htlgmm fits a generalized linear model via generalized method of moments,
The input requires main study and external study.
}
\details{
GWAS for htlgmm.
}
